Add track types and update UART/Motor interfaces
Introduce explicit track types and reference state handling, and adapt UART/Motor interfaces accordingly.

- FY_System.h: add REF system state, reference error flag, fix Refrence -> Reference, add FY_Track enum and FY_Track_t struct (replacing previous uint8_t track fields) and explanatory ASCII diagram.
- Motor/Motor.h: cast SPEED_TO_TIMER results to uint16_t for TIMER_V_MIN/TIMER_V_MAX.
- UART/UART.h: update UART_Context_t to use FY_Track_t* (Track_INFO) and adjust field names/spacing.
- UART/UART.cpp: use systemStatus->error in handleGetError, return Track_INFO in handleGetTrack, add a skeleton implementation and validation checks for handleSetTrack, plus minor formatting fixes.

These changes align track representation across modules and prepare SetTrack handling; further validation/logic is TODO.

+++ FY_Controller_Motor/lib/Motor/Motor.h
@@ -20,8 +20,9 @@
 constexpr uint16_t V_MIN = 1;
 constexpr uint16_t V_MAX = 50;
 
-constexpr uint16_t TIMER_V_MIN = SPEED_TO_TIMER(V_MIN);
-constexpr uint16_t TIMER_V_MAX = SPEED_TO_TIMER(V_MAX);
+constexpr uint16_t TIMER_V_MIN = static_cast<uint16_t>(SPEED_TO_TIMER(V_MIN));
+
+constexpr uint16_t TIMER_V_MAX = static_cast<uint16_t>(SPEED_TO_TIMER(V_MAX));
